+++ src/YamlFilesPathService.php
@@ -10,15 +10,20 @@ use RegexIterator;
 class YamlFilesPathService
 {
     /**
-     * @param string[] $pathToDirs
+     * @param string[] $pathToDirsOrFiles
      * @param string[] $excludedFileMasks
      * @return array
      */
-    public static function getPathToYamlFiles(array $pathToDirs, array $excludedFileMasks = [])
+    public static function getPathToYamlFiles(array $pathToDirsOrFiles, array $excludedFileMasks = [])
     {
         $pathToFiles = [];
-        foreach ($pathToDirs as $pathToDir) {
-            $recursiveDirectoryIterator = new RecursiveDirectoryIterator($pathToDir);
+        foreach ($pathToDirsOrFiles as $pathToDirOrFile) {
+            if (is_file($pathToDirOrFile)) {
+                $pathToFiles[] = $pathToDirOrFile;
+                continue;
+            }
+
+            $recursiveDirectoryIterator = new RecursiveDirectoryIterator($pathToDirOrFile);
             $recursiveIteratorIterator = new RecursiveIteratorIterator($recursiveDirectoryIterator);
             $regexIterator = new RegexIterator($recursiveIteratorIterator, '/^.+\.yml$/i', RecursiveRegexIterator::GET_MATCH);
 
@@ -35,6 +40,8 @@ class YamlFilesPathService
             }
         }
 
-        return $pathToFiles;
+        $pathToFiles = str_replace('\\', '/', $pathToFiles);
+
+        return array_unique($pathToFiles);
     }
 }