+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - option directories changed to required argument
+- to argument for directories you can send file path and directories
 
 ## [1.0.1] - 2017-11-05
 ### Fixed